FloppyGeometry improvements
Added missing XML comments
Adjusted parameter types

+++ TotalImage.IO/DiskGeometries/FloppyGeometry.cs
@@ -3,10 +3,13 @@ using System.Collections.Immutable;
 
 namespace TotalImage.DiskGeometries
 {
+    /// <summary>
+    /// Class that represents a floppy disk geometry/format.
+    /// </summary>
     public class FloppyGeometry
     {
         /// <summary>
-        /// What this?
+        /// Number of density holes.
         /// </summary>
         public byte Hole { get; }
         /// <summary>
@@ -16,43 +19,43 @@ namespace TotalImage.DiskGeometries
         /// <summary>
         /// The data rate.
         /// </summary>
-        public ushort DataRate { get; }
+        public byte DataRate { get; }
         /// <summary>
         /// Encoding type, either FM or MFM.
         /// </summary>
-        public ushort Encoding { get; }
+        public byte Encoding { get; }
         /// <summary>
         /// Revolutions per minute.
         /// </summary>
-        public ushort RPM { get; }
+        public byte RPM { get; }
         /// <summary>
         /// Number of tracks per side.
         /// </summary>
-        public ushort Tracks { get; }
+        public byte Tracks { get; }
         /// <summary>
         /// Number of sectors per track.
         /// </summary>
-        public ushort SPT { get; }
+        public byte SPT { get; }
         /// <summary>
-        /// Number of bytes per sector.
+        /// Number of bytes per sector. Left-shift 128 by this number to get the actual number of bytes.
         /// </summary>
-        public ushort BPS { get; }
+        public byte BPS { get; }
         /// <summary>
         /// Media descriptor byte.
         /// </summary>
-        public ushort MediaDescriptor { get; }
+        public byte MediaDescriptor { get; }
         /// <summary>
         /// Number of sectors per cluster.
         /// </summary>
-        public ushort SPC { get; }
+        public byte SPC { get; }
         /// <summary>
         /// Number of file allocation tables, usually two.
         /// </summary>
-        public ushort NoOfFATs { get; }
+        public byte NoOfFATs { get; }
         /// <summary>
         /// Number of sectors per FAT.
         /// </summary>
-        public ushort SPF { get; }
+        public byte SPF { get; }
         /// <summary>
         /// Number of root directory entries.
         /// </summary>
@@ -60,7 +63,10 @@ namespace TotalImage.DiskGeometries
         /// <summary>
         /// Number of reserved sectors, usually one.
         /// </summary>
-        public ushort ReservedSectors { get; }
+        public byte ReservedSectors { get; }
+        /// <summary>
+        /// Friendly names for known floppy disk geometries.
+        /// </summary>
 
         public enum FriendlyName
         {
@@ -138,8 +144,25 @@ namespace TotalImage.DiskGeometries
             ExtendedDensity
         }
 
-        public FloppyGeometry(byte hole, byte sides, ushort datarate, ushort encoding, ushort rpm, ushort tracks, ushort spt, ushort bps,
-            ushort mediadesc, ushort spc, ushort noofFATs, ushort spf, ushort rootdirentries, ushort reservedsectors)
+        /// <summary>
+        /// Creates a new floppy geometry with provided parameters.
+        /// </summary>
+        /// <param name="hole">Number of density holes, either 0, 1 or 2.</param>
+        /// <param name="sides">Number of sides, either 1 or 2.</param>
+        /// <param name="datarate">Data rate.</param>
+        /// <param name="encoding">Encoding type, either FM or MFM.</param>
+        /// <param name="rpm">Revolution per minute.</param>
+        /// <param name="tracks">Number of tracks per side.</param>
+        /// <param name="spt">Number of sectors per track.</param>
+        /// <param name="bps">Number of bytes per sector.</param>
+        /// <param name="mediadesc">Media descriptor byte.</param>
+        /// <param name="spc">Number of sectors per cluster.</param>
+        /// <param name="noofFATs">Number of file allocation tables.</param>
+        /// <param name="spf">Number of sectors per FAT.</param>
+        /// <param name="rootdirentries">Maximum number of root directory entries.</param>
+        /// <param name="reservedsectors">Number of reserved sectors.</param>
+        public FloppyGeometry(byte hole, byte sides, byte datarate, byte encoding, byte rpm, byte tracks, byte spt, byte bps,
+            byte mediadesc, byte spc, byte noofFATs, byte spf, ushort rootdirentries, byte reservedsectors)
         {
             Hole = hole;
             Sides = sides;
@@ -157,27 +180,29 @@ namespace TotalImage.DiskGeometries
             ReservedSectors = reservedsectors;
         }
 
-        //Key is the friendly name for a geometry/format, value is the FloppyGeometry object containing actual parameters
+        /// <summary>
+        /// List of known floppy disk geometries.
+        /// </summary>
         public static readonly IReadOnlyDictionary<FriendlyName, FloppyGeometry> KnownGeometries = ImmutableDictionary.CreateRange(new Dictionary<FriendlyName, FloppyGeometry>()
         {
             { FriendlyName.DoubleDensity160k, new FloppyGeometry(0, 1, 2, 1, 0, 40,  8, 2, 0xFE, 2, 2,  1,  64, 1) },
             { FriendlyName.DoubleDensity180k, new FloppyGeometry(0, 1, 2, 1, 0, 40,  9, 2, 0xFC, 2, 2,  1,  64, 1) },
-            { FriendlyName.SingleDensity, new FloppyGeometry(0, 1, 2, 1, 0, 77, 26, 0, 0xFE, 4, 2,  6,  68, 1) },
+            { FriendlyName.SingleDensity,     new FloppyGeometry(0, 1, 2, 1, 0, 77, 26, 0, 0xFE, 4, 2,  6,  68, 1) },
             { FriendlyName.DoubleDensity320k, new FloppyGeometry(0, 2, 2, 1, 0, 40,  8, 2, 0xFF, 2, 2,  1, 112, 1) },
             { FriendlyName.DoubleDensity360k, new FloppyGeometry(0, 2, 2, 1, 0, 40,  9, 2, 0xFD, 2, 2,  2, 112, 1) },
-            { FriendlyName.AlphatronicPC16, new FloppyGeometry(0, 2, 2, 1, 0, 40,  5, 3, 0xFF, 2, 2,  1, 128, 1) },
-            { FriendlyName.QuadDensity, new FloppyGeometry(0, 2, 2, 1, 0, 80,  8, 2, 0xFB, 2, 2,  2, 112, 1) },
-            { FriendlyName.Tandy2000, new FloppyGeometry(0, 2, 2, 1, 0, 80,  9, 2, 0xFD, 4, 2,  2, 112, 1) },
-            { FriendlyName.SiemensPCD, new FloppyGeometry(0, 2, 2, 1, 0, 80,  9, 2, 0xF9, 4, 2,  2, 144, 1) },
+            { FriendlyName.AlphatronicPC16,   new FloppyGeometry(0, 2, 2, 1, 0, 40,  5, 3, 0xFF, 2, 2,  1, 128, 1) },
+            { FriendlyName.QuadDensity,       new FloppyGeometry(0, 2, 2, 1, 0, 80,  8, 2, 0xFB, 2, 2,  2, 112, 1) },
+            { FriendlyName.Tandy2000,         new FloppyGeometry(0, 2, 2, 1, 0, 80,  9, 2, 0xFD, 4, 2,  2, 112, 1) },
+            { FriendlyName.SiemensPCD,        new FloppyGeometry(0, 2, 2, 1, 0, 80,  9, 2, 0xF9, 4, 2,  2, 144, 1) },
             { FriendlyName.DoubleDensity720k, new FloppyGeometry(0, 2, 2, 1, 0, 80,  9, 2, 0xF9, 2, 2,  3, 112, 1) },
-            { FriendlyName.Eagle1600, new FloppyGeometry(0, 2, 2, 1, 0, 80,  5, 3, 0xFD, 1, 2,  2, 320, 1) },
-            { FriendlyName.HighDensity1200k, new FloppyGeometry(1, 2, 0, 1, 1, 80, 15, 2, 0xF9, 1, 2,  7, 224, 1) },
-            { FriendlyName.HighDensity1232k, new FloppyGeometry(1, 2, 0, 1, 1, 77,  8, 3, 0xFE, 1, 2,  2, 192, 1) },
-            { FriendlyName.HighDensity1440k, new FloppyGeometry(1, 2, 0, 1, 0, 80, 18, 2, 0xF0, 1, 2,  9, 224, 1) },
-            { FriendlyName.DMF1024, new FloppyGeometry(1, 2, 0, 1, 0, 80, 21, 2, 0xF0, 2, 2,  5,  16, 1) },
-            { FriendlyName.DMF2048, new FloppyGeometry(1, 2, 0, 1, 0, 80, 21, 2, 0xF0, 4, 2,  3,  16, 1) },
-            { FriendlyName.HighDensity1722k, new FloppyGeometry(1, 2, 0, 1, 0, 82, 21, 2, 0xF0, 1, 2, 10, 224, 1) },
-            { FriendlyName.ExtendedDensity, new FloppyGeometry(2, 2, 3, 1, 0, 80, 36, 2, 0xF0, 2, 2,  9, 240, 1) },
+            { FriendlyName.Eagle1600,         new FloppyGeometry(0, 2, 2, 1, 0, 80,  5, 3, 0xFD, 1, 2,  2, 320, 1) },
+            { FriendlyName.HighDensity1200k,  new FloppyGeometry(1, 2, 0, 1, 1, 80, 15, 2, 0xF9, 1, 2,  7, 224, 1) },
+            { FriendlyName.HighDensity1232k,  new FloppyGeometry(1, 2, 0, 1, 1, 77,  8, 3, 0xFE, 1, 2,  2, 192, 1) },
+            { FriendlyName.HighDensity1440k,  new FloppyGeometry(1, 2, 0, 1, 0, 80, 18, 2, 0xF0, 1, 2,  9, 224, 1) },
+            { FriendlyName.DMF1024,           new FloppyGeometry(1, 2, 0, 1, 0, 80, 21, 2, 0xF0, 2, 2,  5,  16, 1) },
+            { FriendlyName.DMF2048,           new FloppyGeometry(1, 2, 0, 1, 0, 80, 21, 2, 0xF0, 4, 2,  3,  16, 1) },
+            { FriendlyName.HighDensity1722k,  new FloppyGeometry(1, 2, 0, 1, 0, 82, 21, 2, 0xF0, 1, 2, 10, 224, 1) },
+            { FriendlyName.ExtendedDensity,   new FloppyGeometry(2, 2, 3, 1, 0, 80, 36, 2, 0xF0, 2, 2,  9, 240, 1) },
         });
     }
 }
